WIP: add kubectl reverse port forwarding
This will add reverse port forwarding feature to kubectl and
kubelet, which means users are able to make a request to a Pod port,
and the request will be forwarded to a local port on the machine running
`kubectl`. More information is described in #20227.

Currently, this PR only covers the `kubectl` part:
A new argument `--reverse` or `-r` is introduced to `port-forward`
subcommand. For example, if you want to open a listening port `8080` on
the Pod that forwards to port `80` on the local machine, you should run
`kubectl port-forward --reverse $POD_NAME 8080:80`.

Please help review if this feature and the command line syntax is
reasonable.

For the kubelet part, I think it would be a little complex to
implement because we should initiate streams between kubectl and kubelet
when new connections come from remote, not local like normal
`port-forward` does. I'll put my proposal later if this feature is
reasonable.